pyproject.toml
[tool.poetry]
name = "telegram-training"
version = "0.1.0"
description = ""
authors = ["Vladislav Vasilev <akiro_jpg@vk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.6"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
isort = "^5.12.0"
aiosqlite = "^0.19.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
